refactor: extract the Priority type to its own file
