+++ src/main/kotlin/com/github/lppedd/cc/completion/providers/FooterValueCompletionProvider.kt
@@ -5,6 +5,7 @@ package com.github.lppedd.cc.completion.providers
 import com.github.lppedd.cc.CCBundle
 import com.github.lppedd.cc.MAX_ITEMS_PER_PROVIDER
 import com.github.lppedd.cc.api.*
+import com.github.lppedd.cc.completion.Priority
 import com.github.lppedd.cc.completion.resultset.ResultSet
 import com.github.lppedd.cc.configuration.CCConfigService
 import com.github.lppedd.cc.lookupElement.CommitFooterLookupElement
