+++ src/main/kotlin/com/github/lppedd/cc/completion/providers/ProviderWrapper.kt
@@ -1,6 +1,7 @@
 package com.github.lppedd.cc.completion.providers
 
 import com.github.lppedd.cc.api.CommitTokenProvider
+import com.github.lppedd.cc.completion.Priority
 import org.jetbrains.annotations.ApiStatus
 
 /**
@@ -9,15 +10,4 @@ import org.jetbrains.annotations.ApiStatus
 @ApiStatus.Internal
 internal interface ProviderWrapper : CommitTokenProvider {
   fun getPriority(): Priority
-}
-
-internal inline class Priority(val value: Int) : Comparable<Priority> {
-  override fun compareTo(other: Priority): Int =
-    value.compareTo(other.value)
-
-  operator fun plus(other: Int): Int =
-    value + other
-
-  operator fun times(times: Int): Int =
-    value * times
 }
